user_service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
psycopg2-binary
bcrypt==4.3.0
python-jose[cryptography]
passlib
pydantic
python-multipart